test: recupera e organiza testes de linha quente para manter simetria
